Added confirmation in FindCustomer

+++ src/components/FindCustomer.jsx
@@ -33,11 +33,14 @@ const FindCustomer = () => {
   }, [gstNo, invoiceNo]);
 
   const handleDelete = async (customerId) => {
-    try {
-      await deleteCustomer(customerId); // Call the service function to delete customer
-      console.log('Customer deleted successfully:', customerId);
-    } catch (error) {
-      console.error('Error deleting customer:', error);
+    const confirmDelete = window.confirm('Are you sure you want to delete the customer?');
+    if(confirmDelete){
+      try {
+        await deleteCustomer(customerId); // Call the service function to delete customer
+        console.log('Customer deleted successfully:', customerId);
+      } catch (error) {
+        console.error('Error deleting customer:', error);
+      }
     }
     navigate('/');
   };
